feat: export defaultConnectors
