+++ packages/connectkit/src/defaultConfig.ts
@@ -7,17 +7,11 @@ import {
 } from 'wagmi';
 import { Chain, mainnet, polygon, optimism, arbitrum } from 'wagmi/chains';
 
-import { MetaMaskConnector } from 'wagmi/connectors/metaMask';
-import { WalletConnectConnector } from 'wagmi/connectors/walletConnect';
-import { WalletConnectLegacyConnector } from 'wagmi/connectors/walletConnectLegacy';
-import { CoinbaseWalletConnector } from 'wagmi/connectors/coinbaseWallet';
-import { SafeConnector } from 'wagmi/connectors/safe';
-import { InjectedConnector } from 'wagmi/connectors/injected';
-
 import { alchemyProvider } from 'wagmi/providers/alchemy';
 import { infuraProvider } from 'wagmi/providers/infura';
 import { jsonRpcProvider } from 'wagmi/providers/jsonRpc';
 import { publicProvider } from 'wagmi/providers/public';
+import defaultConnectors from './defaultConnectors';
 
 let globalAppName: string;
 let globalAppIcon: string;
@@ -26,17 +20,6 @@ export const getAppName = () => globalAppName;
 export const getAppIcon = () => globalAppIcon;
 
 const defaultChains = [mainnet, polygon, optimism, arbitrum];
-
-type DefaultConnectorsProps = {
-  chains?: Chain[];
-  app: {
-    name: string;
-    icon?: string;
-    description?: string;
-    url?: string;
-  };
-  walletConnectProjectId?: string;
-};
 
 type DefaultConfigProps = {
   appName: string;
@@ -61,87 +44,6 @@ type ConnectKitClientProps = {
   connectors?: Connector[];
   publicClient: PublicClient;
   webSocketPublicClient?: WebSocketPublicClient;
-};
-
-const getDefaultConnectors = ({
-  chains,
-  app,
-  walletConnectProjectId,
-}: DefaultConnectorsProps) => {
-  const hasAllAppData = app.name && app.icon && app.description && app.url;
-  const shouldUseSafeConnector =
-    !(typeof window === 'undefined') && window?.parent !== window;
-
-  let connectors: Connector[] = [];
-
-  // If we're in an iframe, include the SafeConnector
-  if (shouldUseSafeConnector) {
-    connectors = [
-      ...connectors,
-      new SafeConnector({
-        chains,
-        options: {
-          allowedDomains: [/gnosis-safe.io$/, /app.safe.global$/],
-          debug: false,
-        },
-      }),
-    ];
-  }
-
-  // Add the rest of the connectors
-  connectors = [
-    ...connectors,
-    new MetaMaskConnector({
-      chains,
-      options: {
-        shimDisconnect: true,
-        UNSTABLE_shimOnConnectSelectAccount: true,
-      },
-    }),
-    new CoinbaseWalletConnector({
-      chains,
-      options: {
-        appName: app.name,
-        headlessMode: true,
-      },
-    }),
-    walletConnectProjectId
-      ? new WalletConnectConnector({
-          chains,
-          options: {
-            showQrModal: false,
-            projectId: walletConnectProjectId,
-            metadata: hasAllAppData
-              ? {
-                  name: app.name,
-                  description: app.description!,
-                  url: app.url!,
-                  icons: [app.icon!],
-                }
-              : undefined,
-          },
-        })
-      : new WalletConnectLegacyConnector({
-          chains,
-          options: {
-            qrcode: false,
-          },
-        }),
-    new InjectedConnector({
-      chains,
-      options: {
-        shimDisconnect: true,
-        name: (detectedName) =>
-          `Injected (${
-            typeof detectedName === 'string'
-              ? detectedName
-              : detectedName.join(', ')
-          })`,
-      },
-    }),
-  ];
-
-  return connectors;
 };
 
 const defaultConfig = ({
@@ -189,7 +91,7 @@ const defaultConfig = ({
     autoConnect,
     connectors:
       connectors ??
-      getDefaultConnectors({
+      defaultConnectors({
         chains: configuredChains,
         app: {
           name: appName,
